refactor: slam function
