+++ src/kf.rs
@@ -1,4 +1,4 @@
-pub mod agent;
-pub mod kf;
+pub mod kfagent;
+pub mod kfilter;
 // pub mod robot;
 pub mod sensor;
