+++ static/robotics_archi.drawio.svg:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=https://app.diagrams.net/
